+++ CHANGELOG.md
@@ -11,6 +11,16 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 
 ### Changed
 
+-
+
+### Fixed
+
+-
+
+# [2.4.10] - 2026-01-19
+
+### Changed
+
 - Changed total variables to total variable references in scan statistics output.
 
 ### Fixed